backend/requirements.txt
altgraph==0.17.2
auto-py-to-exe==2.10.0
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
Eel==0.12.4
future==0.18.2
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.1
idna==3.2
numpy==1.21.2
opencv-python==4.5.3.56
pefile==2021.9.3
playsound==1.2.2
pycparser==2.20
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
python-dotenv==0.19.0
pywin32==301
pywin32-ctypes==0.2.0
requests==2.26.0
urllib3==1.26.6
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
Flask
flask-cors
pyinstaller